Extract image details into a separate component
This works around an issue where Jest and / or
React Testing Library cannot be used to test
components that generate metadata

See https://github.com/vercel/next.js/issues/47299
